[Detail]
	Target: 6455a150fae41e9736d9440d7c36fa7e19fc02c1

MTK-Commit-Id: be087bb0eac18562d6f27698ba757518c211c455

Feature: Others
CR-Id: ALPS05605552

+++ net/mac80211/mesh_hwmp.c
@@ -355,7 +355,7 @@ static u32 airtime_link_metric_get(struct ieee80211_local *local,
 	 */
 	tx_time = (device_constant + 10 * test_frame_len / rate);
 	estimated_retx = ((1 << (2 * ARITH_SHIFT)) / (s_unit - err));
-	result = (tx_time * estimated_retx) >> (2 * ARITH_SHIFT);
+	result = ((u64)tx_time * estimated_retx) >> (2 * ARITH_SHIFT);
 	return (u32)result;
 }
